+++ app/schemas/admin/auth.py
@@ -1,9 +1,5 @@
-from datetime import datetime
-from typing import Optional
-
 from pydantic import Field
 
-from app.enums import Status
 from app.schemas.base import BaseSchema
 
 
@@ -14,20 +10,3 @@ class LoginIn(BaseSchema):
 
 class LoginOut(BaseSchema):
     access_token: str
-
-
-class AuthorIn(BaseSchema):
-    name: str
-    description: Optional[str]
-    image_url: Optional[str]
-    status: Optional[Status]
-
-
-class AuthorOut(BaseSchema):
-    id: int
-    date_created: datetime
-    date_modified: datetime
-    name: str
-    description: Optional[str]
-    image_url: Optional[str]
-    status: Status
