+++ app/views/admin/__init__.py
@@ -6,6 +6,7 @@ from starlette.exceptions import HTTPException as StarletteHTTPException
 
 from app.database.database import SessionLocal
 from app.views.admin.endpoints.auth import router as auth_router
+from app.views.admin.endpoints.author import router as author_router  # Author router eklenmeli
 from app.views.admin.endpoints.book import router as book_router
 from app.views.admin.endpoints.category import router as category_router
 
@@ -17,6 +18,7 @@ app = FastAPI(
 app.include_router(auth_router, prefix="/auth", tags=["Auth"])
 app.include_router(book_router, prefix="/book", tags=["Book"])
 app.include_router(category_router, prefix="/category", tags=["Category"])
+app.include_router(author_router, prefix="/author", tags=["Author"])
 
 
 @app.exception_handler(StarletteHTTPException)
